[COMPLETE] LP-1674 OTHER Send emails to prompt learners to view & download certificates [CHANGE] please signals to the more appropriate file.
