+++ lms/djangoapps/bulk_email/signals.py
@@ -26,17 +26,3 @@ def force_optout_all(sender, **kwargs):  # pylint: disable=unused-argument
 
     for enrollment in CourseEnrollment.objects.filter(user=user):
         Optout.objects.get_or_create(user=user, course_id=enrollment.course.id)
-
-
-@receiver(USER_CERTIFICATE_DOWNLOADABLE)
-def send_email_user_certificate_downloadable(sender, first_name, display_name, certificate_reverse_url, user_email,
-                                             *args, **kwargs):
-    template = MandrillClient.DOWNLOAD_CERTIFICATE
-
-    context = dict(first_name=first_name, course_name=display_name,
-                   certificate_url=certificate_reverse_url)
-    try:
-        MandrillClient().send_mail(template, user_email, context)
-    except Exception as e:
-        # Mandrill errors are thrown as exceptions
-        log.error('Unable to send email for USER_CERTIFICATE_DOWNLOADABLE signal: %s - %s' % (e.__class__, e))
